[FLINK-1328] [api-breaking][java-api][scala-api][optimizer] Reworked semantic annotations for functions. - Renamed constantField annotations to forwardedFields annotation - Forwarded fields can be defined for (nested) tuples, Pojos, case classes - Added semantic function information to example programs
This closes #311

+++ flink-core/src/main/java/org/apache/flink/api/common/operators/Ordering.java
@@ -144,7 +144,7 @@ public class Ordering {
 		}
 		
 		for (int i = 0; i < this.indexes.size(); i++) {
-			if (this.indexes.get(i).intValue() != otherOrdering.indexes.get(i).intValue()) {
+			if (this.indexes.get(i) != otherOrdering.indexes.get(i)) {
 				return false;
 			}
